src/modules/cars/useCase/listCars/ListAvailableCarsUseCase.ts
import { ICarsRepository } from "@modules/cars/repositories/ICarsRepository";
import { inject, injectable } from "tsyringe";

interface IRequest {
  category_id?: string;
  brand?: string;
  name?: string;
}

@injectable()
export class ListAvailableCarsUseCase {
  constructor(
    @inject("CarsRepository")
    private carsRepository: ICarsRepository
  ) {}

  async execute({ brand, category_id, name }: IRequest) {
    const cars = this.carsRepository.listAllAvailable(brand, category_id, name);
    return cars;
  }
}
